+++ cmd/hidden_lake/service/internal/handler/config_settings.go
@@ -5,6 +5,7 @@ import (
 	"net/http"
 
 	"github.com/number571/go-peer/cmd/hidden_lake/service/internal/config"
+	pkg_config "github.com/number571/go-peer/cmd/hidden_lake/service/pkg/config"
 	pkg_settings "github.com/number571/go-peer/cmd/hidden_lake/service/pkg/settings"
 	"github.com/number571/go-peer/internal/api"
 	http_logger "github.com/number571/go-peer/internal/logger/http"
@@ -27,13 +28,16 @@ func HandleConfigSettingsAPI(pWrapper config.IWrapper, pLogger logger.ILogger, p
 			pLogger.PushInfo(logBuilder.WithMessage(http_logger.CLogSuccess))
 
 			sett := pWrapper.GetConfig().GetSettings()
-			api.Response(pW, http.StatusOK, config.SConfigSettings{
-				FMessageSizeBytes:   sett.GetMessageSizeBytes(),
-				FWorkSizeBits:       sett.GetWorkSizeBits(),
-				FQueuePeriodMS:      sett.GetQueuePeriodMS(),
-				FKeySizeBits:        sett.GetKeySizeBits(),
-				FLimitVoidSizeBytes: sett.GetLimitVoidSizeBytes(),
-				FNetworkKey:         sett.GetNetworkKey(),
+			api.Response(pW, http.StatusOK, pkg_config.SConfigSettings{
+				SConfigSettings: config.SConfigSettings{
+					FMessageSizeBytes:   sett.GetMessageSizeBytes(),
+					FWorkSizeBits:       sett.GetWorkSizeBits(),
+					FQueuePeriodMS:      sett.GetQueuePeriodMS(),
+					FKeySizeBits:        sett.GetKeySizeBits(),
+					FLimitVoidSizeBytes: sett.GetLimitVoidSizeBytes(),
+					FNetworkKey:         sett.GetNetworkKey(),
+				},
+				FLimitMessageSizeBytes: pNode.GetMessageQueue().GetClient().GetMessageLimit(),
 			})
 			return