+++ internal/base64/base64_test.go
@@ -1,0 +1,5 @@
+package base64
+
+import "testing"
+
+func TestNothing(t *testing.T) {}